LayoutTests/http/tests/inspector/stacktraces/csp-setTimeout-warning-contains-stacktrace-expected.txt

CONSOLE MESSAGE: Refused to evaluate script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

This test attempts to evaluate script via setTimeout. The resulting console error should contain a stack trace tied to line 11.

Message[0]: csp-setTimeout-warning-contains-stacktrace.html:11 Refused to evaluate script because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline'".

Stack Trace:

    0) http://127.0.0.1:8000/inspector/stacktraces/csp-setTimeout-warning-contains-stacktrace.html / thisTest / 11 / 0
